+++ docs/user_guide/half_float_casting.py
@@ -1,7 +1,11 @@
 # Copyright (c) 2020 Graphcore Ltd. All rights reserved.
 # pragma pylint: disable=function-redefined
+import sys
 import torch
 import poptorch
+
+if not poptorch.hasMLIRSupportOnPlatform():
+    sys.exit(0)
 
 
 # Cases in which casting resolves to the correct type
@@ -55,6 +59,8 @@ assert native_model(float16_tensor, float16_tensor).dtype == torch.float32
 assert native_model(float32_tensor, float16_tensor).dtype == torch.float32
 
 opts = poptorch.Options()
+# Important: this is only needed for traceModel(True)
+opts.Jit.traceModel(True)
 opts.Precision.halfFloatCasting(
     poptorch.HalfFloatCastingBehavior.HalfUpcastToFloat)
 
@@ -65,4 +71,13 @@ assert poptorch_model(float16_tensor, float16_tensor).dtype == torch.float16
 # This incorrectly results in a float 16 tensor
 poptorch_model = poptorch.inferenceModel(native_model, opts)
 assert poptorch_model(float32_tensor, float16_tensor).dtype == torch.float16
+
+# UPDATE: with the new default of traceModel(False) PopTorch now matches the native behaviour
+poptorch_model = poptorch.inferenceModel(native_model)
+assert poptorch_model(float16_tensor, float16_tensor).dtype == native_model(
+    float16_tensor, float16_tensor).dtype
+
+poptorch_model = poptorch.inferenceModel(native_model)
+assert poptorch_model(float32_tensor, float16_tensor).dtype == native_model(
+    float32_tensor, float16_tensor).dtype
 # incorrect_cast_end